# packages/api-client/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/finance/screener:
    get:
      operationId: getScreener
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            default: 20
            type: number
        - name: sortBy
          required: false
          in: query
          description: Column to sort by
          schema:
            type: string
        - name: sortOrder
          required: false
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: filters
          required: false
          in: query
          description: JSON-encoded map of filter key to filter value
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTickerDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/tickers:
    get:
      operationId: getTickers
      parameters:
        - name: tickers
          required: true
          in: query
          description: Comma-separated list of ticker symbols
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TickerDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/screener/filters/options:
    get:
      operationId: getScreenerFilterOptions
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenerFilterOptionsDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/screener/filters/tickers:
    get:
      operationId: getScreenerTickerOptions
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Fuzzy search on ticker or company name
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTickerOptionDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/status:
    get:
      operationId: getSyncStatus
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatusDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/history:
    get:
      operationId: getSyncHistoryList
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            minimum: 10
            default: 10
            type: number
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - running
              - success
              - partial_success
              - failed
              - timeout
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSyncHistoryDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/history/{id}:
    get:
      operationId: getSyncHistoryDetail
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncHistoryDetailDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync:
    post:
      operationId: syncScreener
      parameters:
        - name: markets
          required: false
          in: query
          description: >-
            Comma-separated list of markets to scope the sync to; omit to
            sync every market
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/static:
    post:
      operationId: syncStatic
      parameters: []
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/fundamental:
    post:
      operationId: syncFundamental
      parameters: []
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/compound:
    post:
      operationId: syncCompound
      parameters: []
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/sync/technical:
    post:
      operationId: syncTechnical
      parameters: []
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{isin}/sync:
    post:
      operationId: syncSingleTicker
      parameters:
        - name: isin
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{isin}/sync/static:
    post:
      operationId: syncSingleTickerStatic
      parameters:
        - name: isin
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{isin}/sync/fundamental:
    post:
      operationId: syncSingleTickerFundamental
      parameters:
        - name: isin
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{isin}/sync/compound:
    post:
      operationId: syncSingleTickerCompound
      parameters:
        - name: isin
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{isin}/sync/technical:
    post:
      operationId: syncSingleTickerTechnical
      parameters:
        - name: isin
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}:
    get:
      operationId: getTicker
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/fundamental:
    get:
      operationId: getTickerFundamental
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundamentalTickerDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/financial-history:
    get:
      operationId: getTickerFinancialHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialHistoryDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/earnings-history:
    get:
      operationId: getTickerEarningsHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EarningsHistoryDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/altman-history:
    get:
      operationId: getTickerAltmanHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AltmanHistoryDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/chart:
    get:
      operationId: getTickerChart
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: window
          required: true
          in: query
          schema:
            type: string
            enum:
              - 5m
              - 1h
              - 1d
              - 1wk
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerChartDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/market-hours:
    get:
      operationId: getTickerMarketHours
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketHoursDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/tickers/hidden:
    get:
      operationId: getHiddenTickers
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Fuzzy search on isin or ticker
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedHiddenTickerDto'
      security:
        - cookie: []
      tags:
        - Finance
  /api/finance/ticker/{id}/unhide:
    post:
      operationId: unhideTicker
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      tags:
        - Finance
...
info:
  title: TraderTavern API
  description: The TraderTavern API documentation
  version: '0.1'
  contact: {}
...
servers: []
components:
  schemas:
...
    TickerDto:
      type: object
      properties:
        isin:
          type: string
        ticker:
          type: string
        companyName:
          type: string
        sector:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        marketCap:
          type: number
          nullable: true
        peRatio:
          type: number
          nullable: true
        price:
          type: number
          nullable: true
        country:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
          description: Basic description of the company
        employees:
          type: number
          nullable: true
        fiscalYearEnd:
          format: date-time
          type: string
          nullable: true
        mostRecentQuarter:
          format: date-time
          type: string
          nullable: true
        market:
          type: string
          nullable: true
          description: Display name of the market the ticker trades on
        currency:
          type: string
          nullable: true
          description: ISO 4217 currency code the ticker is priced in (e.g. USD, EUR)
        changePercent:
          type: number
          nullable: true
          description: >-
            Change since the last completed session close, in percent (e.g. 1.23
            for +1.23%). While the market is open this compares the last two
            completed closes; once the market has closed for the day it compares
            today's close to yesterday's.
        changePercent5d:
          type: number
          nullable: true
          description: 5 day change, in percent
        changePercent1w:
          type: number
          nullable: true
          description: 1 week change, in percent
        changePercent1m:
          type: number
          nullable: true
          description: 1 month change, in percent
        changePercent3m:
          type: number
          nullable: true
          description: 3 month change, in percent
        changePercent6m:
          type: number
          nullable: true
          description: 6 month change, in percent
        changePercentYtd:
          type: number
          nullable: true
          description: Year-to-date change, in percent
        changePercent1y:
          type: number
          nullable: true
          description: 1 year change, in percent
        rsi14:
          type: number
          nullable: true
          description: RSI(14)
        macd:
          type: number
          nullable: true
          description: MACD(12,26,9) line
        macdSignal:
          type: number
          nullable: true
          description: MACD(12,26,9) signal line
        macdHistogram:
          type: number
          nullable: true
          description: MACD(12,26,9) histogram
        bbUpper:
          type: number
          nullable: true
          description: Bollinger Band(20,2) upper
        bbMiddle:
          type: number
          nullable: true
          description: Bollinger Band(20,2) middle
        bbLower:
          type: number
          nullable: true
          description: Bollinger Band(20,2) lower
        bbWidth:
          type: number
          nullable: true
          description: Bollinger Band width, in percent
        atr14:
          type: number
          nullable: true
          description: ATR(14)
        volumeRatio20d:
          type: number
          nullable: true
          description: 20 day volume ratio (latest volume / 20d avg volume)
        refreshedAt:
          format: date-time
          type: string
          nullable: true
          description: When the underlying technical data was last refreshed
        website:
          type: string
          nullable: true
          description: Company website URL
        logoUrl:
          type: string
          nullable: true
          description: Company logo image URL
        psRatio:
          type: number
          nullable: true
        forwardPE:
          type: number
          nullable: true
        pegRatio:
          type: number
          nullable: true
        evToEbitda:
          type: number
          nullable: true
        evToRevenue:
          type: number
          nullable: true
        priceToBook:
          type: number
          nullable: true
        epsTrailing:
          type: number
          nullable: true
        epsForward:
          type: number
          nullable: true
        enterpriseValue:
          type: number
          nullable: true
        fiftyTwoWeekHigh:
          type: number
          nullable: true
        fiftyTwoWeekLow:
          type: number
          nullable: true
        revenue:
          type: number
          nullable: true
          description: Revenue (TTM)
        grossProfit:
          type: number
          nullable: true
        netIncome:
          type: number
          nullable: true
          description: Net income (TTM)
        revenuePerShare:
          type: number
          nullable: true
        ebitda:
          type: number
          nullable: true
        grossMargin:
          type: number
          nullable: true
          description: In percent
        operatingMargin:
          type: number
          nullable: true
          description: In percent
        ebitdaMargin:
          type: number
          nullable: true
          description: In percent
        profitMargin:
          type: number
          nullable: true
          description: In percent
        returnOnEquity:
          type: number
          nullable: true
          description: In percent
        returnOnAssets:
          type: number
          nullable: true
          description: In percent
        revenueGrowth:
          type: number
          nullable: true
          description: In percent
        operatingCashflow:
          type: number
          nullable: true
        freeCashflow:
          type: number
          nullable: true
          description: Free cash flow (TTM)
        capex:
          type: number
          nullable: true
          description: Capital expenditure (TTM)
        totalDebt:
          type: number
          nullable: true
        totalCash:
          type: number
          nullable: true
        debtToEquity:
          type: number
          nullable: true
        currentRatio:
          type: number
          nullable: true
        quickRatio:
          type: number
          nullable: true
        bookValuePerShare:
          type: number
          nullable: true
        dividendYield:
          type: number
          nullable: true
          description: In percent
        payoutRatio:
          type: number
          nullable: true
          description: In percent
        fiveYearAvgDividendYield:
          type: number
          nullable: true
          description: In percent
        exDividendDate:
          format: date-time
          type: string
          nullable: true
        analystRating:
          type: string
          nullable: true
          description: e.g. "buy", "hold"
        analystTargetMean:
          type: number
          nullable: true
        analystTargetLow:
          type: number
          nullable: true
        analystTargetHigh:
          type: number
          nullable: true
        analystCount:
          type: number
          nullable: true
        sharesOutstanding:
          type: number
          nullable: true
        floatShares:
          type: number
          nullable: true
        insidersPercent:
          type: number
          nullable: true
          description: In percent
        institutionsPercent:
          type: number
          nullable: true
          description: In percent
        piotroskiScore:
          type: number
          nullable: true
          description: Piotroski F-Score (0-9)
        altmanZScore:
          type: number
          nullable: true
          description: >-
            Altman Z-Score. > 2.99 safe zone, 1.81-2.99 grey zone, < 1.81
            distress zone
        sma50:
          type: number
          nullable: true
        sma200:
          type: number
          nullable: true
        beta:
          type: number
          nullable: true
        avgVolume10d:
          type: number
          nullable: true
          description: 10 day average volume
      required:
        - isin
        - ticker
        - companyName
        - sector
        - industry
        - marketCap
        - peRatio
        - price
        - country
        - description
        - employees
        - fiscalYearEnd
        - mostRecentQuarter
        - market
        - currency
        - changePercent
        - changePercent5d
        - changePercent1w
        - changePercent1m
        - changePercent3m
        - changePercent6m
        - changePercentYtd
        - changePercent1y
        - rsi14
        - macd
        - macdSignal
        - macdHistogram
        - bbUpper
        - bbMiddle
        - bbLower
        - bbWidth
        - atr14
        - volumeRatio20d
        - refreshedAt
        - website
        - logoUrl
        - psRatio
        - forwardPE
        - pegRatio
        - evToEbitda
        - evToRevenue
        - priceToBook
        - epsTrailing
        - epsForward
        - enterpriseValue
        - fiftyTwoWeekHigh
        - fiftyTwoWeekLow
        - revenue
        - grossProfit
        - netIncome
        - revenuePerShare
        - ebitda
        - grossMargin
        - operatingMargin
        - ebitdaMargin
        - profitMargin
        - returnOnEquity
        - returnOnAssets
        - revenueGrowth
        - operatingCashflow
        - freeCashflow
        - capex
        - totalDebt
        - totalCash
        - debtToEquity
        - currentRatio
        - quickRatio
        - bookValuePerShare
        - dividendYield
        - payoutRatio
        - fiveYearAvgDividendYield
        - exDividendDate
        - analystRating
        - analystTargetMean
        - analystTargetLow
        - analystTargetHigh
        - analystCount
        - sharesOutstanding
        - floatShares
        - insidersPercent
        - institutionsPercent
        - piotroskiScore
        - altmanZScore
        - sma50
        - sma200
        - beta
        - avgVolume10d
    PaginatedTickerDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TickerDto'
        meta:
          $ref: '#/components/schemas/PaginationMetaDto'
      required:
        - data
        - meta
    ScreenerFilterOptionsDto:
      type: object
      properties:
        sectors:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        countries:
          type: array
          items:
            type: string
        markets:
          type: array
          items:
            type: string
        currencies:
          type: array
          items:
            type: string
        analystRatings:
          type: array
          items:
            type: string
      required:
        - sectors
        - industries
        - countries
        - markets
        - currencies
        - analystRatings
...
    PaginatedTickerOptionDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TickerOptionDto'
        meta:
          $ref: '#/components/schemas/PaginationMetaDto'
      required:
        - data
        - meta
    SyncStatusDto:
      type: object
      properties:
        lastSyncDate:
          format: date-time
          type: string
          nullable: true
      required:
        - lastSyncDate
...
    SyncHistoryDetailDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - auto
            - manual
        kind:
          type: string
          enum:
            - ticker
            - static
            - fundamental
            - compound
            - technical
            - single_ticker
        status:
          type: string
          enum:
            - running
            - success
            - partial_success
            - failed
            - timeout
        syncDate:
          format: date-time
          type: string
        market:
          type: string
          nullable: true
        marketLabel:
          type: string
          nullable: true
        tickerCount:
          type: number
        succeededCount:
          type: number
        failedCount:
          type: number
        triggeredByUserId:
          type: string
          nullable: true
        triggeredByUsername:
          type: string
          nullable: true
        startedAt:
          format: date-time
          type: string
        finishedAt:
          format: date-time
          type: string
          nullable: true
        tickers:
          type: array
          items:
            $ref: '#/components/schemas/SyncHistoryTickerDto'
        generalError:
          type: string
          nullable: true
      required:
        - id
        - type
        - kind
        - status
        - syncDate
        - market
        - marketLabel
        - tickerCount
        - succeededCount
        - failedCount
        - triggeredByUserId
        - triggeredByUsername
        - startedAt
        - finishedAt
        - tickers
        - generalError
    PaginatedSyncHistoryDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SyncHistoryListItemDto'
        meta:
          $ref: '#/components/schemas/PaginationMetaDto'
      required:
        - data
        - meta
...